fix(app): last ui improvements
